Rewrite txn family tutorial for changes to code and repo structure
Rewrite txn family tutorial to account for code changes related
to check_valid function, as well as to not use the git branch
method of changing the code in the tuturial. Instead, use
subdirectories named after the steps, and a simple bash script
that changes the PYTHONPATH environment variable so that the
python code used by the validator and user changes to reflect
the step the user is on in the tutorial.