Rewrite Ubuntu 16.04 installation section
Previously, we asked the user to execute commands as root.  Generally,
guides such as this one use sudo in the instructions so that the user
does not need to have a root shell.

This uses add-apt-repository command instead of appending to the
sources.list directly with tee. This provides a friendlier sudo command.

Separates out repo management from the actual installation command for
clarity.  Also provides the user with a method to view which sawtooth
packages are installed.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>