Move hero-bg.jpg to dedicated images folder
Previously, this file was part of the examples/ directory. The examples
directory will be removed later, but this image will still be used, so
it is moved out of the examples directory.

This also de-duplicates hero-bg.jpg, which also appeared in docs;
instead, this now just links to the images/ version.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>