Add redirects for renamed files
This commit adds Markdown "redirect files" for each renamed smart contract spec file.

It adds back the previous file with a redirect_to directive, as a separate commit
(in the hope that Markdown will keep the file history with the renamed file).

This change will prevent "404 File not found" messages for anyone using the old link to the Grid transaction family specs.

Signed-off-by: Anne Chenette <chenette@bitwise.io>