Add separate proc for Creating Splinter Circuits
This new how-to procedure expands the original circuit-creation steps (from
"Running Grid on Splinter") with prerequisites, clarifying info, corrections,
and subheadings for navigation. Plus a new section on determining the fully
qualified service ID and setting GRID_SERVICE_ID to this value.

NOTE: Although the intro, nav, and linking content is new, the example commands
are exactly the same as in "Running Grid on Splinter". The only differences:
- Added output for "splinter circuit propose"
- Replaced a pair of angle brackets with curly braces
- Add example of "splinter circuit show $CIRCUIT_ID" for the accepted circuit,
  so that this section shows both the list and show subcommands.

Also note: The "Next Steps" section in this procedure links to another new
topic, "Using Grid Features", which was added in a different PR.

Signed-off-by: Anne Chenette <chenette@bitwise.io>