New Using Grid Features doc topic
This new topic summarizes the recommended order of using Grid features
(org/agents first, then schema, then product) and the prerequisites (2+ grid
nodes; connected by a circuit for Grid on Splinter).

As separate Grid how-to procedures are added, this topic will link to those
procedures.

Each of the separate topics will have a "Next Steps" section that links back
to this topic. This approach isolates the impact of link changes when a new
how-to procedure is added; only this file will need to be updated with the new
link.

Signed-off-by: Anne Chenette <chenette@bitwise.io>