Update release pages to link to release branch
This change updates the links used in the `Releases` pages for Grid v0.1
and v0.2 to update the GitHub link to point to the specific release
branch. This change also adds a note to these release pages to specify
that the released version exists on a branch, rather than just the core
repository.

Signed-off-by: Shannyn Telander <telander@bitwise.io>