Clarify step in Creating Locations
Explicity adds a step in Creating Locations doc to create a locations admin
and assign that role to the agent. Previously, this step was mentioned in
prereqs, but this change includes it in the directions as it is a critical
step. This reduces the chance that a reader misses this step and encounters
errors.

This commit also updates numbering.

Signed-off-by: Chris Eckhardt <eckhardt@bitwise.io>