Update future rest api documentation
This updates the endpoints for the future rest api docs to remove
explicit references to Griddle. This is intended to avoid confusion and
allow these docs to evolve with the project. This also includes updating
the url of the docs to remove references to Griddle. A list of changes
to the API spec has also been added to the API landing page.

Signed-off-by: Davey Newhall <newhall@bitwise.io>