Fix permissions for using Grid Location
This change fixes the line in the `Creating Locations` guide used to
create the role with permissions to create, update, and delete a
location. Previously, the list of permissions was broken up by a `\` to
allow it to span multiple lines. However, this was creating a space in
the `location::can-delete-location` permission that disabled Grid from
being able to recognize this permission. This change removes the back
slash and keeps the permissions on one line, which still allows for
linting to pass and does not cause a space to occur in the permissions
string.

This change fixes the creating location guide across v0.2, v0.3 and v0.4
docs.

Signed-off-by: Shannyn Telander <telander@bitwise.io>