Fix payload schema for agents
This change updates the payload schema for agents to separate the struct
used to create an agent payload. These are separated because the `Agent`
struct for the GET agent endpoints will only need to be serialized to be
returned to the user. The payload resources will need to be deserialized
from the request body. Therefore, these objects will be implemented
separately.

Signed-off-by: Shannyn Telander <telander@bitwise.io>