Fix payload for purchase order versions
This change updates the purchase order version schema to be separated
for the GET endpoints and the batch-creating endpoints. The GET
endpoints will use a structure that is serializable to be returned to
the client. The batch-creating endpoints, however, will need to use the
structure to deserialize the client's request. Therefore, in
implementation, these objects will be separated.

Signed-off-by: Shannyn Telander <telander@bitwise.io>