Add polling calls to DLT event monitor document
The DLT event monitor document previously called for batch updates to
happen in real-time as they came in from the DLT stream. Unfortunately,
the current implementation of the DLT event stream does not include
sufficient information to allow a useful batch status update. This
change proposes an alternate solution: The DLT event monitor uses events
as a cue that a batch was updated. It then calls out to the polling
monitor, which gathers the associated batch information.

Signed-off-by: Lee Bradley <bradley@bitwise.io>