Update walkthrough of creating a splinter circuit
This change updates the walkthrough that explains how to make a Splinter
circuit using the Grid docker-compose example. Grid's Splinter
dependency was updated to v0.6, which changed the API slightly. Mainly,
this removes the `--key` and `--url` options and updates the command
output.

This change also adds a copyright to the top of the file.

Signed-off-by: Shannyn Telander <telander@bitwise.io>